Removing inappropriate test
This test isn't doing what's needed. The need is to ensure that the Engine's forms create form inputs which provide necessary data to this controller's action for creation of Answerables. This test as currently written sets up its own form inputs, which may be different from those generated by the engine. In other words, this test should stay in the Portal, where it was written.