Refactoring finding the @page in controller
We need a matched set of LightweightActivity and InteractivePage. If we have the Activity ID, we should use that first, because it will generate an error when we try to find the page. If we don't have it, we find whatever page is specified, and use that LightweightActivity, which may or may not be the intended one.